+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Alias shell as window, Glimmer::UI::CustomShell as Glimmer::UI::CustomWindow, on_shell_xyz events as on_window_xyz events, and :shell_trim SWT style as :window_trim SWT Glimmer custom style
 - `bind` option of `async_exec` or `sync_exec` (e.g. `bind(model, :property, async_exec: true)`) to perform GUI updates from another thread safely
 - Show progress bar for Mandelbrot calculation
+- Mandelbrot Cores menu options to switch the number of CPU cores being used for calculation on the next zoom calculation cycle
 
 ### 4.18.4.9